apps/generation-service/src/services/imagen_service.py
"""Google Imagen 3 image generation service."""

import asyncio
from typing import List, Optional
from google import genai

from .config import settings


class ImagenService:
    """Service for generating images using Google Imagen 3."""
    
    def __init__(self):
        self.client = genai.Client(
            api_key=settings.google_ai_api_key,
            vertexai=True,
            project=settings.vertex_project_id,
            location=settings.vertex_location,
        )
        self.model = settings.imagen_model
    
    async def generate(
        self,
        prompt: str,
        variations: int = 1,
        resolution: str = "1024x1024",
        style: str = "photorealistic",
    ) -> List[dict]:
        """Generate images from text prompt."""
        
        # TODO: Implement actual Imagen 3 API call
        
        await asyncio.sleep(2)  # Simulate processing
        
        results = []
        for i in range(variations):
            results.append({
                "image_url": f"https://example.com/image_{i}.png",
                "resolution": resolution,
            })
        
        return results
    
    async def upscale(
        self,
        image_url: str,
        scale: int = 2,
    ) -> dict:
        """Upscale image."""
        
        # TODO: Implement upscaling
        
        return {"image_url": "https://example.com/upscaled.png"}


imagen_service = ImagenService()
