apps/generation-service/pyproject.toml
[project]
name = "generation-service"
version = "1.0.0"
description = "NOVA AI Generation Service - Python FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-generativeai>=0.3.2",
    "google-cloud-storage>=2.14.0",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "pino-python>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
